Remove unused dependency react-email
This package is not imported anywhere and has a lot of transitional
runtime dependencies. Some of these are clearly build/development
dependencies (e.g. swc, webpack) and they are even propagated into
NextJS standalone builds, making it much bigger than it needs to be.

Note that there're also @react-email/components and @react-email/render
- unlike "react-email", these dependencies are used.